requirements-test.txt
# test runs requirements (versions we'll be testing against) - automatically updated by requires.io
coverage==4.1 # pytest-cov
pylama==7.0.9
pyflakes==1.2.3
pep257==0.7.0
pep8==1.7.0
mccabe==0.5.0
pyroma==2.0.2
mirakuru==0.8.2
path.py==8.2.1
port-for==0.3.1
pymongo==3.3.0
pytest==2.8.7 #  rq.filter: !=2.9.0,!=2.9.1,!=2.9.2
